First version of az-ops-sb
This is a cli for management and operational aspects for Azure Service Bus.
The initial scope is to be able to visualize dead letters and requeue them.

This commit is the same version as the 0.0.1 that was published to nuget
but with some cleanup when it comes to naming of namespaces and noisy git
history that wouldn't benefit any.